Cleanup & Refactoring
Refactored some code to prep for the second part
