+++ Day 3 Mull It Over/Shared/TokenType.cs
@@ -1,0 +1,8 @@
+﻿namespace AdventOfCode_2024.DayThree.Shared;
+
+public enum TokenType
+{
+    MulFunction,
+    DoFunction,
+    DontFunction
+}
